requirements.txt
langchain
openai
anthropic
chromadb
pinecone-client
weasyprint
pandas
numpy
yfinance
python-jose
pyjwt
frappe-ui
